backend/requirements.txt
# Web Framework
fastapi~=0.111.0
uvicorn[standard]~=0.30.0
python-multipart~=0.0.9

# Configuration & Validation
pydantic~=2.7.0
pydantic-settings~=2.2.0
email-validator~=2.1.0

# Database & ORM (Relational - Postgres)
sqlalchemy[asyncio]~=2.0.30
asyncpg~=0.29.0
alembic~=1.13.0

# Graph Database (Neo4j)
neo4j~=5.20.0

# Auth & Security
bcrypt~=4.1.0
passlib[bcrypt]~=1.7.4
pyjwt~=2.8.0

# Background Jobs & Caching
celery~=5.4.0
redis~=5.0.0

# Document Processing
pymupdf~=1.24.4
python-docx~=1.1.0
tiktoken~=0.7.0

# NLP - Keyword / Concept Extraction
spacy~=3.7.4

# AI & Utilities
google-genai~=0.5.0
python-dotenv~=1.0.0
tenacity~=8.3.0
httpx~=0.27.0

# Dev & Testing
pytest~=8.2.0
pytest-asyncio~=0.23.0
ruff~=0.4.0
black~=24.4.0
isort~=5.13.0
mypy~=1.10.0

# Production & Monitoring
gunicorn~=22.0.0
python-json-logger~=2.0.7
sentry-sdk[fastapi]~=2.1.1
